ffmpeg: Support for external tracks and partial ytdl support
This makes external subtitles work and it makes youtube-dl streams
somewhat work. It won't work with multi-part streams (edl) which
require some manifest file and I can't figure out how to make ffmpeg
understand them. Presumably however, to get it to work would require
the script to parse the segment files and timings itself and build the
ffmpeg command from that, which might be too complicated/reinventing-
the-wheel-ish. So, that can be put on ice.